struct関連が複雑なので、ここだけ分離他の箇所のリファクタリングの邪魔になっている
